+++ src/components/DetailedOffer/CoachCard.tsx
@@ -7,6 +7,11 @@ interface CoachCardProps {
   coachId?: number;
 }
 
+/**
+ * This component fetches and displays the name and the picture of the selected offer's coach
+ * @param coachId the ID of the coach to fetch
+ * @returns The coach card
+ */
 const CoachCard = ({ coachId }: CoachCardProps) => {
   const { isLoading, data: coach } = useQuery<Coach>(`coach ${coachId}`, () =>
     getCoach(coachId ?? 0)
@@ -19,10 +24,10 @@ const CoachCard = ({ coachId }: CoachCardProps) => {
       ) : (
         <div className="flex flex-row items-center space-x-4 px-3">
           <Avatar
-            size="xl"
-            src={coach?.user?.photo}
             alt="Coach"
             className="rounded-[3rem]"
+            size="xl"
+            src={coach?.user?.photo}
           />
           <div className="flex space-x-2 ml-2 text-lg xs:text-xl sm:text-3xl items-center line-clamp-1">
             <span className="font-bold">Coach : </span>
